added in calling and search

+++ client/src/components/Home/index.js
@@ -4,7 +4,7 @@ import axios from 'axios';
 // import { getBooks } from '../../store/actions/book_actions';
 // import {RowGenerator, GenreateRowsWithBlocks} from '../../utils/helpers';
 import Patient from '../../components/Patient'
-import { getPatients, deletePatient } from '../../store/actions/patient_actions'
+import { getPatients, deletePatient, patientCalled } from '../../store/actions/patient_actions'
 import { addSchedule } from '../../store/actions/scheduled_actions'
 import {connect} from 'react-redux';
 
@@ -22,7 +22,7 @@ import * as Twilio from 'twilio-client';
 class Home extends Component {
 
     state = {
-        patients: []
+        patientSearch: ''
     }
 
     componentDidMount(){
@@ -39,23 +39,25 @@ class Home extends Component {
         
     }
 
-    // call=()=>{
-    //     axios.get('/api/calls/token').then(res =>{
-    //         console.log(res)
-    //         Twilio.Device.setup(res.data)
-    //     })
-
-              
-    // }
+    call=(patient)=>{
+        this.props.dispatch(patientCalled(this.props.patients.collection[patient]))
+        this.props.dispatch(getPatients())
+        this.props.dispatch(getPatients())      
+    }
 
     scheduledPateint=(key)=>{
 
 
-        alert('Confirm Person has been scheduled')
-        console.log(this.props.patients.collection[key])
-        this.props.dispatch(addSchedule(this.props.patients.collection[key]))
-        this.props.dispatch(deletePatient(this.props.patients.collection[key]))
-        this.props.dispatch(getPatients())
+        // alert('Confirm Person has been scheduled')
+        let answer = window.confirm("Are you sure this person is in the Schedule")
+        if(answer){
+            console.log(this.props.patients.collection[key])
+            this.props.dispatch(addSchedule(this.props.patients.collection[key]))
+            this.props.dispatch(deletePatient(this.props.patients.collection[key]))
+            this.props.dispatch(getPatients())
+            this.props.dispatch(getPatients())
+        }
+        console.log('no')
 
 
     }
@@ -64,7 +66,14 @@ class Home extends Component {
     showPatients =(patients) => {
         if(patients.collection){
             return(
-                patients.collection.map((item, index)=>(
+                patients.collection.filter(val=>{
+                    if(this.state.patientSearch == ''){
+                        return val
+                    }else if (val.name.toLowerCase().includes(this.state.patientSearch.toLocaleLowerCase())){
+                        return val
+                    }
+
+                }).map((item, index)=>(
                     <TableRow key={index}>
                         <TableCell>
                             {item.name}
@@ -73,8 +82,8 @@ class Home extends Component {
                         <TableCell>{item.location}</TableCell>
                         <TableCell>{item.phoneNumber}</TableCell>
                         <TableCell>{item.referralDate}</TableCell>
-                        <TableCell><button>Called</button></TableCell>
-                        <TableCell><button onClick={() => this.scheduledPateint(index)}>Scheduled</button></TableCell>
+                        <TableCell><button className="Login-button button1" onClick={() => this.call(index)}><div>{item.called}</div>Called</button></TableCell>
+                        <TableCell><button className="Login-button button1" onClick={() => this.scheduledPateint(index)}>Scheduled</button></TableCell>
                     </TableRow>
                 ))
             )
@@ -94,7 +103,7 @@ class Home extends Component {
                 <hr/>
                 <div className="App-container">
                 <div className='index-container'>
-                    <input type="text" name="search" placeholder="Search Patient"/>
+                    <input type="text" name="search" placeholder="Search Patient" onChange={e=>this.setState({patientSearch: e.target.value})}/>
                 </div>
 
                 <div className='index-container'>
